requirements.txt
Django==3.2.9
sqlalchemy>=1.0.14
django-environ==0.8.1
psycopg2-binary==2.9.2
gunicorn==20.1.0
drf-yasg>=1.20.0
djangorestframework==3.15.1
celery==5.4.0
